custom ui refactor
