+++ public/index.html
@@ -170,8 +170,8 @@
     </div>
 
     <script type="module" src="/js/autocomplete-setup.js"></script>
-    <script type="module" src="/js/custom-dialog.js"></script>
-    <script type="module" src="/js/custom-toast.js"></script>
+    <script type="module" src="/js/custom-ui/dialog.js"></script>
+    <script type="module" src="/js/custom-ui/toast.js"></script>
     <script type="module" src="/js/main.js"></script>
   </body>
 </html>
