+++ public/js/custom-ui/gallery.js
